seven commit

+++ order-service/src/main/java/microservice/order_service/order/OrderController.java
@@ -27,7 +27,7 @@ public class OrderController {
 
     @GetMapping("/{order-id}")
     public ResponseEntity<OrderResponse> findById(
-            @PathVariable("order-id") Integer orderId
+            @PathVariable("order-id") Long orderId
     ) {
         return ResponseEntity.ok(this.service.findById(orderId));
     }
