+++ order-service/src/main/java/microservice/order_service/orderline/OrderLineRequest.java
@@ -1,7 +1,7 @@
 package microservice.order_service.orderline;
 
 public record OrderLineRequest(
-        Integer orderId,
+        Long orderId,
         Integer productId,
         double quantity
 ) {
